Update README.md
Nitesh Synergy (Nitesh) is building an AI tool that eliminates the need for multiple layers of hierarchy. The system will operate with just a few essential roles: Developer, QA, CEO, and a single level of Controller who acts as a bridge between the CEO and the Tech Lead.

We are removing the need for roles like Senior Manager, Manager, Agile Manager, and other similar positions. These roles can effectively be replaced by NiteshSynergy.ai.